install llamafile 0.9.0.
llamafile now implements its own server for chat/completion requests, in addition to embedding. We are not yet using it, beceause it is not clear which advantages it has for API users. A current deal breaker is that it does not support assigning an alias for a model. We're using that to easily provide a user-friendly name e.g. when using ollama weights.